Split out ListenerManager and friends.
This includes the messages it uses, as well as CometListener and
LiftActorJWithListenerManager.
