+++ web/webkit/src/main/scala/net/liftweb/http/CometActor.scala
@@ -207,212 +207,10 @@ trait LiftCometActor extends TypedActor[Any, Any] with ForwardableActor[Any, Any
 
 object CurrentCometActor extends ThreadGlobal[LiftCometActor]
 
-object AddAListener {
-  def apply(who: SimpleActor[Any]) = new AddAListener(who, {
-    case _ => true
-  })
-}
-
-/**
- * This is a message class for use with ListenerManager and CometListener
- * instances. The use of the shouldUpdate function is deprecated, and
- * should instead be handled by the message processing partial functions
- * on the CometListener instances themselves.
- *
- * @see CometListener
- * @see ListenerManager
- */
-case class AddAListener(who: SimpleActor[Any], shouldUpdate: PartialFunction[Any, Boolean])
-
-/**
- * This is a message class for use with ListenerManager and CometListener
- * instances.
- *
- * @see CometListener
- * @see ListenerManager
- */
-case class RemoveAListener(who: SimpleActor[Any])
-
-
-object ListenerManager {
-  type ActorTest = (SimpleActor[Any], PartialFunction[Any, Boolean])
-}
-
-
-/**
- * This trait manages a set of Actors in a publish/subscribe pattern. When you extend your Actor with
- * this trait, you automatically get handling for sending messages out to all subscribed Actors. Simply
- * override the high-, medium-, or lowPriority handlers to do your message processing. When you want to update
- * all subscribers, just call the updateListeners method. The createUpdate method is used to generate
- * the message that you want sent to all subscribers.
- *
- * Note that the AddAListener and RemoveAListener messages (for subscription control) are processed
- * after any highPriority or mediumPriority messages are processed, so take care to avoid overly
- * broad matches in those handlers that might consume internal messages.
- *
- * For example, you could write a simple service to provide clock ticks using the following code:
- *
- * <pre name="code" class="scala">
- * case object Tick
- *
- * object Ticker extends ListenerManager with LiftActor {
- *   import net.liftweb.util.ActorPing
- *
- *   // Set up the initial tick
- *   ActorPing.schedule(this, Tick, 1000L)
- *
- *   // This is a placeholder, since we're only interested
- *   // in Ticks
- *   def createUpdate = "Registered"
- *
- *   override def mediumPriority = {
- *     case Tick => {
- *       sendListenersMessage(Tick)
- *       ActorPing.schedule(this, Tick, 1000L)
- * }
- * }
- * }
- * </pre>
- *
- * A client CometActor could look like:
- *
- * <pre name="code" class="scala">
- * class CometClock extends CometListener {
- *   val registerWith = Ticker
- *
- *   ... handling code ...
- * }
- * </pre>
- *
- * @see CometListener
- *
- */
-trait ListenerManager {
-  self: SimpleActor[Any] =>
-
-  import ListenerManager._
-
-  /**
-   * This is the list of all registered actors
-   */
-  private var listeners: List[ActorTest] = Nil
-
-  protected def messageHandler: PartialFunction[Any, Unit] =
-    highPriority orElse mediumPriority orElse
-      listenerService orElse lowPriority
-
-  protected def listenerService: PartialFunction[Any, Unit] = {
-    case AddAListener(who, wantsMessage) =>
-      val pair = (who, wantsMessage)
-      listeners ::= pair
-
-      updateListeners(pair :: Nil)
-
-    case RemoveAListener(who) =>
-      listeners = listeners.filter(_._1 ne who)
-      if (listeners.isEmpty) {
-        onListenersListEmptied()
-      }
-  }
-
-  /**
-   * Called after RemoveAListener-message is processed and no more listeners exist.
-   * Default does nothing.
-   */
-  protected def onListenersListEmptied() {
-  }
-
-  /**
-   * Update the listeners with the message generated by createUpdate
-   */
-  protected def updateListeners(listeners: List[ActorTest] = listeners) {
-    val update = createUpdate
-
-    listeners foreach {
-      case (who, wantsMessage) if wantsMessage.isDefinedAt(update) && wantsMessage(update) =>
-        who ! update
-    }
-  }
-
-  /**
-   * Send a message we create to all of the listeners. Note that with this
-   * invocation the createUpdate method is not used.
-   */
-  protected def sendListenersMessage(msg: Any) {
-    listeners foreach (_._1 ! msg)
-  }
-
-  /**
-   * This method is called when the <pre>updateListeners()</pre> method
-   * needs a message to send to subscribed Actors. In particular, createUpdate
-   * is used to create the first message that a newly subscribed CometListener
-   * will receive.
-   */
-  protected def createUpdate: Any
-
-  /**
-   * Override this method to process high priority messages. Note:
-   * <b>you must not process messages with a wildcard (match all)</b>, since
-   * this will intercept the messages used for subscription control.
-   */
-  protected def highPriority: PartialFunction[Any, Unit] = Map.empty
-
-  /**
-   * Override this method to process medium priority messages. See
-   * the highPriority method for an important note on wildcard
-   * processing.
-   *
-   * @see #highPriority
-   */
-  protected def mediumPriority: PartialFunction[Any, Unit] = Map.empty
-
-  /**
-   * Override this method to process low priority messages.
-   */
-  protected def lowPriority: PartialFunction[Any, Unit] = Map.empty
-}
-
 /**
  * A LiftActorJ with ListenerManager.  Subclass this class
  * to get a Java-usable LiftActorJ with ListenerManager
  */
-abstract class LiftActorJWithListenerManager extends LiftActorJ with ListenerManager {
-  protected override def messageHandler: PartialFunction[Any, Unit] =
-    highPriority orElse mediumPriority orElse
-      listenerService orElse lowPriority orElse _messageHandler
-}
-
-/**
- * This trait adds functionality to automatically register with a given
- * Actor using AddAListener and RemoveAListener control messages. The most
- * typical usage would be to register with an instance of ListenerManager.
- * You will need to provide a def/val for the <pre>registerWith</pre> member
- * to control which Actor to connect to.
- *
- * See ListenerManager for a complete example.
- *
- * @see ListenerManager
- */
-trait CometListener extends BaseCometActor {
-  self: BaseCometActor =>
-
-  /**
-   * This controls which Actor to register with for updates. Typically
-   * this will be an instance of ListenerActor, although you can provide
-   * your own subscription handling on top of any SimpleActor.
-   */
-  protected def registerWith: SimpleActor[Any]
-
-  abstract override protected def localSetup() {
-    registerWith ! AddAListener(this,  { case _ => true })
-    super.localSetup()
-  }
-
-  abstract override protected def localShutdown() {
-    registerWith ! RemoveAListener(this)
-    super.localShutdown()
-  }
-}
 
 /**
  * Subclass from this class if you're in Java-land
@@ -422,14 +220,6 @@ abstract class CometActorJ extends LiftActorJ with CometActor {
 
   override def lowPriority = _messageHandler
 
-}
-
-/**
- * Subclass from this class if you want a CometActorJ with
- * CometListeners
- */
-abstract class CometActorJWithCometListener extends CometActorJ with CometListener {
-  override def lowPriority = _messageHandler
 }
 
 trait CometActor extends BaseCometActor {
